fix: snap, test rename
